requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.2.11
Flask-Mail==0.9.0
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.7
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.1
bitstring==3.1.3
blinker==1.3
cffi==0.8.6
itsdangerous==0.24
pefile==1.2.10-114
pluginbase==0.3
pycparser==2.10
pygeoip==0.3.1
pysqlite==2.6.3
python-dateutil==2.2
python-magic==0.4.6
pytz==2014.4
requests==2.3.0
simplejson==3.6.2
six==1.7.3
speaklater==1.3
ssdeep==3.1
urllib3==1.9
uuid==1.30
virustotal==1.0.3
wsgiref==0.1.2
